fix(st): cast clk_get_rate return when setting console
The return of clk_get_rate() is an unsigned long. But for console, it
will never be above 4GHz, so it is safe to cast it to uint32_t, before
passing the parameter to set_console().

This corrects the warning:
plat/st/common/stm32mp_common.c:303:13: warning: implicit conversion
 loses integer precision: 'unsigned long' to 'uint32_t' (aka 'unsigned
 int')
      [-Wshorten-64-to-32]
  303 |         clk_rate =
		 	clk_get_rate((unsigned long)dt_uart_info.clock);
      |                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ plat/st/common/stm32mp_common.c
@@ -300,7 +300,7 @@ int stm32mp_uart_console_setup(void)
 
 	reset_uart((uint32_t)dt_uart_info.reset);
 
-	clk_rate = clk_get_rate((unsigned long)dt_uart_info.clock);
+	clk_rate = (uint32_t)clk_get_rate((unsigned long)dt_uart_info.clock);
 #endif
 
 	set_console(dt_uart_info.base, clk_rate);
